options_scanner/requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
firebase-admin>=6.2.0
python-dotenv>=1.0.0
pytz>=2023.3
